+++ app/decorators/dnd2024_character/classes/fighter_decorator.rb
@@ -20,6 +20,12 @@ module Dnd2024Character
       def spells_slots
         @spells_slots ||= ::Dnd2024Character::ClassDecorateWrapper::EMPTY_SPELL_SLOTS[class_level]
       end
+
+      private
+
+      def class_level
+        @class_level ||= classes['fighter']
+      end
     end
   end
 end